requirements.txt
# Copyright 2017 Google Inc. All rights reserved.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

# google-auth (without `six`)
google-auth==1.2.1
cachetools==2.0.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
rsa==3.4.2

# `requests`
requests==2.18.4
certifi==2017.11.5
chardet==3.0.4
idna==2.6
urllib3==1.22

# `Flask`
click==6.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.12.2
Flask==0.12.2
